# Speed optimization for optimal parser