Finally found a way to integrate
a new speed optimization for LZ4 optimal parser, named Chain Swap,
into the generic match finder,
thus supporting all search scenarios (extDict and CDict).

Results are pretty good : 
benchmark on `entropy`, `gcc-7`, highest compression mode `-12` :

| Name | -12 (dev) | -12 (this patch) | Improvement |
| --- | --- | --- | --- |
| calgary.tar | 6.2 MB/s | 16.2 MB/s | +160 % |
| enwik7 | 15.6 MB/s | 18.7 MB/s | +20 % |
| silesia.tar | 5.8 MB/s | 13.8 MB/s | +140 % |

Performance improvement vary depending on file, but good news is,
slowest files see the biggest improvements.

Benefits are mostly present at level 12.
Level 11 still see some benefit, but much lower, in the ~20% range.
Level 10 see a strange reversal, with same speed or slightly slower, but better compression ratio.
That's because, instead of reducing the nb of searches, which are now pretty low, the optimization makes each search more useful, but also more complex.

This optimization is not enabled at levels 9 and below,
as the trade-off between better ratio vs slower speed becomes less interesting.